+++ client/package.json
@@ -3,6 +3,7 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
+		"cookie": "^0.4.0",
 		"dotenv": "^8.2.0",
 		"jsonwebtoken": "^8.5.1",
 		"prop-types": "^15.7.2",
